+++ log_code/code.go
@@ -8,9 +8,8 @@ const (
 	WarnHttpServerShutdown = 602
 	ErrorHttpServerListen  = 603
 
-	FatalLocalConfig = 610
-	ErrorClientGrpc  = 606
-	ErrorClientHttp  = 611
+	ErrorClientGrpc = 606
+	ErrorClientHttp = 611
 
 	ErrorClientJournal = 608
 	ErrorClientRedis   = 609