+++ routing/routing.go
@@ -5,20 +5,24 @@ import (
 )
 
 var (
-	InnerAddressMap = make(map[string]bool)
+	InnerMethods = make(map[string]bool)
+	AllMethods   = make(map[string]bool)
 )
 
 func InitRoutes(configs structure.RoutingConfig) {
+	newAddressMap := make(map[string]bool)
 	newInnerAddressMap := make(map[string]bool)
 	for _, backend := range configs {
 		if backend.Address.IP == "" || backend.Address.Port == "" || len(backend.Endpoints) == 0 {
 			continue
 		}
 		for _, v := range backend.Endpoints {
+			newAddressMap[v.Path] = true
 			if v.Inner {
 				newInnerAddressMap[v.Path] = true
 			}
 		}
 	}
-	InnerAddressMap = newInnerAddressMap
+	AllMethods = newAddressMap
+	InnerMethods = newInnerAddressMap
 }