+++ CHANGELOG.md
@@ -1,2 +1,9 @@
-### v2.0.0
+## v2.0.0
 * update to new isp-lib & config service
+## 1.1.1
+* add x-instance-identity header
+## 1.1.0
+* add token verification via JWT
+* add accounting for applications
+* add optional cache for application tokens
+